+++ src/gql/resolvers/FormResolver.php
@@ -3,8 +3,8 @@ namespace verbb\formie\gql\resolvers;
 
 use verbb\formie\elements\Form;
 use verbb\formie\helpers\Gql as GqlHelper;
-use verbb\formie\helpers\Table;
 
+use craft\db\Table;
 use craft\elements\db\ElementQuery;
 use craft\elements\ElementCollection;
 use craft\gql\base\ElementResolver;
@@ -38,7 +38,7 @@ class FormResolver extends ElementResolver
         }
 
         if (!GqlHelper::canSchema('formieForms.all')) {
-            $query->andWhere(['in', 'elements.id', array_values(Db::idsByUids(Table::FORMIE_FORMS, $pairs['formieForms']))]);
+            $query->andWhere(['in', 'elements.id', array_values(Db::idsByUids(Table::ELEMENTS, $pairs['formieForms']))]);
         }
 
         return $query;